requirements.txt
scanpy>=1.10.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
umap-learn>=0.5.3
faiss-cpu>=1.7.4
flask>=3.0.0
plotly>=5.15.0
h5py>=3.8.0
openpyxl>=3.1.0
xlrd>=2.0.1
loompy>=3.0.7
werkzeug>=3.0.0
